+++ CHANGELOG.md
@@ -5,7 +5,21 @@ All notable changes to `powerdns-php` will be documented in this file.
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
 ## Unreleased
-[Compare v3.1.0 - Unreleased](https://github.com/exonet/powerdns-php/compare/v3.1.0...develop)
+[Compare v3.2.0 - Unreleased](https://github.com/exonet/powerdns-php/compare/v3.2.0...develop)
+
+## [v3.2.0](https://github.com/exonet/powerdns-php/releases/tag/v3.2.0) - 2021-05-06
+[Compare v3.1.0 - v3.2.0](https://github.com/exonet/powerdns-php/compare/v3.1.0...v3.2.0)
+### Added
+- Added a method `unsetNsec3param()` to unset the nsec3param for a zone.
+- Unit tests for enabling, disabling and setting DNSSEC.
+- It is now possible to set/get resource records to a zone resource. See `examples/new_domain_from_zone_resource.php` for an example.
+
+### Changed
+- Calling `setNsec3Param()` with value `null` will unset the `nsec3param`.
+- The internals of getting ResourceSets. Previously a GET request was made to get the resource sets, now the zone resource is being used which already contains all resource sets.
+
+### Fixed
+- Updated the stable composer.phar URL to properly run tests.
 
 ## [v3.1.0](https://github.com/exonet/powerdns-php/releases/tag/v3.1.0) - 2021-02-01
 [Compare v3.0.0 - v3.1.0](https://github.com/exonet/powerdns-php/compare/v2.0.0...v3.1.0)